feat: Kl = u64, Kw = u32 again

+++ m68k/test/loop.c
@@ -1,7 +1,7 @@
-unsigned long
-loop(unsigned long i)
+unsigned
+loop(unsigned i)
 {
-	unsigned long x;
+	unsigned x;
 	for (; x < i; ++x);
 	return x;
 }
